renames job schema.ts to job-form-validations
