+++ frontend/apps/web/app/(mgmt)/[account]/jobs/util.ts
@@ -85,7 +85,7 @@ import {
   SingleTableSchemaFormValues,
   SubsetFormValues,
   WorkflowSettingsSchema,
-} from '../new/job/schema';
+} from '../new/job/job-form-validations';
 
 type GetConnectionById = (id: string) => Connection | undefined;
