+++ frontend/apps/web/app/(mgmt)/[account]/new/job/job-form-validations.ts
